feat: Lines 클래스 추가
